Save scores to the database
This is the main feature that makes the osu! server actually feel like a proper osu! server.
+ Add database operations for scores
* Fix some places I forgot to change float to double for pp and accuracy
+ Add grade calculator for osu!standard
* Move performance points calculation to somewhere that makes sense
+ Load a player's top 500 plays from the database when they connect to allow for realtime pp calculations as they play
* Prevent a player from connecting to the server twice or with multiple sessions

+++ OsuServer/API/ScoreReport.cs
@@ -47,8 +47,8 @@ namespace OsuServer.API
                 UpdatedValue("rankedScore", _oldScore.TotalScore, _newScore.TotalScore),
                 UpdatedValue("totalScore", _oldScore.TotalScore, _newScore.TotalScore),
                 UpdatedValue("maxCombo", _oldScore.Combo, _newScore.Combo),
-                UpdatedValue("accuracy", _oldScore.Accuracy * 100f, _newScore.Accuracy * 100f),
-                UpdatedValue("pp", (float) Math.Round(_oldScore.PerformancePoints), (float) Math.Round(_newScore.PerformancePoints)),
+                UpdatedValue("accuracy", (float)_oldScore.Accuracy * 100f, (float)_newScore.Accuracy * 100f),
+                UpdatedValue("pp", (float)Math.Round(_oldScore.PerformancePoints), (float)Math.Round(_newScore.PerformancePoints)),
                 $"onlineScoreId:{_bancho.Scores.GetByChecksum(scoreChecksum)}",
                 "\n",
                 $"chartId:overall",
@@ -58,8 +58,8 @@ namespace OsuServer.API
                 UpdatedValue("rankedScore", _oldStats.RankedScore, _newStats.RankedScore),
                 UpdatedValue("totalScore", _oldStats.TotalScore, _newStats.TotalScore),
                 UpdatedValue("maxCombo", _oldStats.MaxCombo, _newStats.MaxCombo),
-                UpdatedValue("accuracy", (float) _oldStats.Accuracy * 100f, (float) _newStats.Accuracy * 100f),
-                UpdatedValue("pp", (float) Math.Round(_oldStats.PP), (float) Math.Round(_newStats.PP)),
+                UpdatedValue("accuracy", (float)_oldStats.Accuracy * 100f, (float)_newStats.Accuracy * 100f),
+                UpdatedValue("pp", (float)Math.Round(_oldStats.PP), (float)Math.Round(_newStats.PP)),
                 $"achievements-new:" // TODO: achievements
             ];
             return string.Join("|", report);
